+++ Tests/Test_SetY_MultiLayer/Test_Numerical_Int/test_b2_radial_odes.py
@@ -60,6 +60,41 @@ def test_derivatives_solid_static():
     for y_i in range(6):
         assert type(solid_derivative[y_i]) in [complex, np.complex128]
 
+def test_derivatives_solid_static_incomp():
+    """ Tests the calculation of radial ODE for a solid layer using the static and incompressible assumption """
+
+    # Initialize 6 fake radial function solutions
+    radial_funcs = (
+            10. + 1.j,
+            10. + 1.j,
+            10. + 1.j,
+            10. + 1.j,
+            10. + 1.j,
+            10. + 1.j
+        )
+
+    # Test for l=2
+    solid_derivative = static_solid_ode(
+        radius_array_to_use[2], radial_funcs,
+        radius_array_to_use, shear_array, bulk_array,
+        density_array, gravity_array, order_l=2, incompressible=True
+        )
+    assert type(solid_derivative) == tuple
+    assert len(solid_derivative) == 6
+    for y_i in range(6):
+        assert type(solid_derivative[y_i]) in [complex, np.complex128]
+
+    # Test for l=3
+    solid_derivative = static_solid_ode(
+        radius_array_to_use[2], radial_funcs,
+        radius_array_to_use, shear_array, bulk_array,
+        density_array, gravity_array, order_l=3, incompressible=True
+        )
+    assert type(solid_derivative) == tuple
+    assert len(solid_derivative) == 6
+    for y_i in range(6):
+        assert type(solid_derivative[y_i]) in [complex, np.complex128]
+
 
 def test_derivatives_solid_dynamic():
     """ Tests the calculation of radial derivatives for a solid layer using the dynamic assumption """
@@ -96,6 +131,40 @@ def test_derivatives_solid_dynamic():
     for y_i in range(6):
         assert type(solid_derivative[y_i]) in [complex, np.complex128]
 
+def test_derivatives_solid_dynamic_incomp():
+    """ Tests the calculation of radial derivatives for a solid layer using the dynamic and incompressible assumption """
+
+    # Initialize 6 fake radial function solutions
+    radial_funcs = (
+        10. + 1.j,
+        10. + 1.j,
+        10. + 1.j,
+        10. + 1.j,
+        10. + 1.j,
+        10. + 1.j
+        )
+
+    # Test for l=2
+    solid_derivative = dynamic_solid_ode(
+        radius_array_to_use[2], radial_funcs,
+        radius_array_to_use, shear_array, bulk_array,
+        density_array, gravity_array, frequency, order_l=2, incompressible=True
+        )
+    assert type(solid_derivative) == tuple
+    assert len(solid_derivative) == 6
+    for y_i in range(6):
+        assert type(solid_derivative[y_i]) in [complex, np.complex128]
+
+    # Test for l=3
+    solid_derivative = dynamic_solid_ode(
+        radius_array_to_use[2], radial_funcs,
+        radius_array_to_use, shear_array, bulk_array,
+        density_array, gravity_array, frequency, order_l=3, incompressible=True
+        )
+    assert type(solid_derivative) == tuple
+    assert len(solid_derivative) == 6
+    for y_i in range(6):
+        assert type(solid_derivative[y_i]) in [complex, np.complex128]
 
 def test_derivatives_liquid_static():
     """ Tests the calculation of radial derivatives for a liquid layer using the static assumption """
@@ -128,6 +197,36 @@ def test_derivatives_liquid_static():
     for y_i in range(2):
         assert type(liquid_derivative[y_i]) in [complex, np.complex128]
 
+def test_derivatives_liquid_static_incomp():
+    """ Tests the calculation of radial derivatives for a liquid layer using the static and incompressible assumption """
+
+    # Initialize 2 fake radial function solutions
+    radial_funcs = (
+        10. + 1.j,
+        10. + 1.j
+        )
+
+    # Test for l=2
+    liquid_derivative = static_liquid_ode(
+        radius_array_to_use[2], radial_funcs,
+        radius_array_to_use, density_array, gravity_array,
+        order_l=2, incompressible=True
+        )
+    assert type(liquid_derivative) == tuple
+    assert len(liquid_derivative) == 2
+    for y_i in range(2):
+        assert type(liquid_derivative[y_i]) in [complex, np.complex128]
+
+    # Test for l=3
+    liquid_derivative = static_liquid_ode(
+        radius_array_to_use[2], radial_funcs,
+        radius_array_to_use, density_array, gravity_array,
+        order_l=3, incompressible=True
+        )
+    assert type(liquid_derivative) == tuple
+    assert len(liquid_derivative) == 2
+    for y_i in range(2):
+        assert type(liquid_derivative[y_i]) in [complex, np.complex128]
 
 def test_derivatives_liquid_dynamic():
     """ Tests the calculation of radial derivatives for a liquid layer using the dynamic assumption """
@@ -156,6 +255,39 @@ def test_derivatives_liquid_dynamic():
         radius_array_to_use[2], radial_funcs,
         radius_array_to_use, bulk_array,
         density_array, gravity_array, frequency, order_l=3
+        )
+    assert type(liquid_derivative) == tuple
+    assert len(liquid_derivative) == 4
+    for y_i in range(4):
+        assert type(liquid_derivative[y_i]) in [complex, np.complex128]
+
+def test_derivatives_liquid_dynamic_incomp():
+    """ Tests the calculation of radial derivatives for a liquid layer using the dynamic and incompressible assumption """
+
+    # Initialize 4 fake radial function solutions
+    radial_funcs = (
+        10. + 1.j,
+        10. + 1.j,
+        10. + 1.j,
+        10. + 1.j
+        )
+
+    # Test for l=2
+    liquid_derivative = dynamic_liquid_ode(
+        radius_array_to_use[2], radial_funcs,
+        radius_array_to_use, bulk_array, density_array,
+        gravity_array, frequency, order_l=2, incompressible=True
+        )
+    assert type(liquid_derivative) == tuple
+    assert len(liquid_derivative) == 4
+    for y_i in range(4):
+        assert type(liquid_derivative[y_i]) in [complex, np.complex128]
+
+    # Test for l=3
+    liquid_derivative = dynamic_liquid_ode(
+        radius_array_to_use[2], radial_funcs,
+        radius_array_to_use, bulk_array,
+        density_array, gravity_array, frequency, order_l=3, incompressible=True
         )
     assert type(liquid_derivative) == tuple
     assert len(liquid_derivative) == 4