+++ TidalPy/tides/multilayer/numerical_int/derivatives/odes.py
@@ -15,7 +15,10 @@ from typing import Tuple
 import numpy as np
 
 from . import (radial_derivatives_liquid_dynamic, radial_derivatives_liquid_static,
-               radial_derivatives_solid_dynamic, radial_derivatives_solid_static)
+               radial_derivatives_solid_dynamic, radial_derivatives_solid_static,
+               radial_derivatives_liquid_dynamic_incomp, radial_derivatives_liquid_static_incomp,
+               radial_derivatives_solid_dynamic_incomp, radial_derivatives_solid_static_incomp
+               )
 from .....constants import G
 from .....utilities.performance import njit
 from .....utilities.types import FloatArray, NumArray
@@ -27,7 +30,7 @@ def dynamic_solid_ode(
     y_vector: Tuple[NumArray, NumArray, NumArray, NumArray, NumArray, NumArray],
     radii: np.ndarray, shear_moduli: np.ndarray, bulk_moduli: np.ndarray,
     densities: np.ndarray, gravities: np.ndarray, frequency: float,
-    order_l: int = 2, G_to_use: float = G
+    order_l: int = 2, G_to_use: float = G, incompressible: bool = False
     ) -> Tuple[NumArray, NumArray, NumArray, NumArray, NumArray, NumArray]:
     """ A njit-safe radial derivative function for static, solid layers.
 
@@ -54,6 +57,8 @@ def dynamic_solid_ode(
         Tidal harmonic order
     G_to_use : float = G
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
+    incompressible : bool = False
+        If `True`, the incompressible assumption will be used.
 
     Returns
     -------
@@ -64,15 +69,22 @@ def dynamic_solid_ode(
     # These physical parameters are generally given in a coarser resolution than what is required during integration.
     #    Therefore, we interpolate their values at the integrator's requested `radius`.
     shear_modulus = np.interp(radius, radii, shear_moduli)
-    bulk_modulus = np.interp(radius, radii, bulk_moduli)
     density = np.interp(radius, radii, densities)
     gravity = np.interp(radius, radii, gravities)
 
-    solid_dyn_derivatives = \
-        radial_derivatives_solid_dynamic(
-            radius, y_vector, shear_modulus, bulk_modulus, density, gravity, frequency,
-            order_l=order_l, G_to_use=G_to_use
-            )
+    if incompressible:
+        solid_dyn_derivatives = \
+            radial_derivatives_solid_dynamic_incomp(
+                radius, y_vector, shear_modulus, density, gravity, frequency,
+                order_l=order_l, G_to_use=G_to_use
+                )
+    else:
+        bulk_modulus = np.interp(radius, radii, bulk_moduli)
+        solid_dyn_derivatives = \
+            radial_derivatives_solid_dynamic(
+                radius, y_vector, shear_modulus, bulk_modulus, density, gravity, frequency,
+                order_l=order_l, G_to_use=G_to_use
+                )
 
     return solid_dyn_derivatives
 
@@ -82,7 +94,7 @@ def static_solid_ode(
     radius: FloatArray, y_vector: Tuple[NumArray, NumArray, NumArray, NumArray, NumArray, NumArray],
     radii: np.ndarray, shear_moduli: np.ndarray, bulk_moduli: np.ndarray,
     densities: np.ndarray, gravities: np.ndarray,
-    order_l: int = 2, G_to_use: float = G
+    order_l: int = 2, G_to_use: float = G, incompressible: bool = False
     ) -> Tuple[NumArray, NumArray, NumArray, NumArray, NumArray, NumArray]:
     """ A njit-safe radial derivative function for static, solid layers.
 
@@ -107,6 +119,8 @@ def static_solid_ode(
         Tidal harmonic order
     G_to_use : float = G
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
+    incompressible : bool = False
+        If `True`, the incompressible assumption will be used.
 
     Returns
     -------
@@ -117,15 +131,22 @@ def static_solid_ode(
     # These physical parameters are generally given in a coarser resolution than what is required during integration.
     #    Therefore, we interpolate their values at the integrator's requested `radius`.
     shear_modulus = np.interp(radius, radii, shear_moduli)
-    bulk_modulus = np.interp(radius, radii, bulk_moduli)
     density = np.interp(radius, radii, densities)
     gravity = np.interp(radius, radii, gravities)
 
-    solid_static_derivatives = \
-        radial_derivatives_solid_static(
-            radius, y_vector, shear_modulus, bulk_modulus, density, gravity,
-            order_l=order_l, G_to_use=G_to_use
-            )
+    if incompressible:
+        solid_static_derivatives = \
+            radial_derivatives_solid_static_incomp(
+                radius, y_vector, shear_modulus, density, gravity,
+                order_l=order_l, G_to_use=G_to_use
+                )
+    else:
+        bulk_modulus = np.interp(radius, radii, bulk_moduli)
+        solid_static_derivatives = \
+            radial_derivatives_solid_static(
+                radius, y_vector, shear_modulus, bulk_modulus, density, gravity,
+                order_l=order_l, G_to_use=G_to_use
+                )
 
     return solid_static_derivatives
 
@@ -135,7 +156,7 @@ def dynamic_liquid_ode(
     radius: FloatArray, y_vector: Tuple[NumArray, NumArray, NumArray, NumArray],
     radii: np.ndarray, bulk_moduli: np.ndarray,
     densities: np.ndarray, gravities: np.ndarray, frequency: float,
-    order_l: int = 2, G_to_use: float = G
+    order_l: int = 2, G_to_use: float = G, incompressible: bool = False
     ) -> callable:
     """ A njit-safe radial derivative function for dynamic, liquid layers.
 
@@ -160,6 +181,8 @@ def dynamic_liquid_ode(
         Tidal harmonic order
     G_to_use : float = G
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
+    incompressible : bool = False
+        If `True`, the incompressible assumption will be used.
 
     Returns
     -------
@@ -169,14 +192,20 @@ def dynamic_liquid_ode(
 
     # These physical parameters are generally given in a coarser resolution than what is required during integration.
     #    Therefore, we interpolate their values at the integrator's requested `radius`.
-    bulk_modulus = np.interp(radius, radii, bulk_moduli)
     density = np.interp(radius, radii, densities)
     gravity = np.interp(radius, radii, gravities)
 
-    liquid_dyn_derivatives = radial_derivatives_liquid_dynamic(
-        radius, y_vector, bulk_modulus, density, gravity, frequency,
-        order_l=order_l, G_to_use=G_to_use
-        )
+    if incompressible:
+        liquid_dyn_derivatives = radial_derivatives_liquid_dynamic_incomp(
+            radius, y_vector, density, gravity, frequency,
+            order_l=order_l, G_to_use=G_to_use
+            )
+    else:
+        bulk_modulus = np.interp(radius, radii, bulk_moduli)
+        liquid_dyn_derivatives = radial_derivatives_liquid_dynamic(
+            radius, y_vector, bulk_modulus, density, gravity, frequency,
+            order_l=order_l, G_to_use=G_to_use
+            )
 
     return liquid_dyn_derivatives
 
@@ -185,7 +214,7 @@ def dynamic_liquid_ode(
 def static_liquid_ode(
     radius: FloatArray, y_vector: Tuple[NumArray, NumArray],
     radii: np.ndarray, densities: np.ndarray, gravities: np.ndarray,
-    order_l: int = 2, G_to_use: float = G
+    order_l: int = 2, G_to_use: float = G, incompressible: bool = False
     ) -> Tuple[NumArray, NumArray]:
     """ A njit-safe radial derivative function for static, liquid layers.
 
@@ -206,6 +235,8 @@ def static_liquid_ode(
         Tidal harmonic order
     G_to_use : float = G
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
+    incompressible : bool = False
+        If `True`, the incompressible assumption will be used.
 
     Returns
     -------
@@ -218,9 +249,15 @@ def static_liquid_ode(
     density = np.interp(radius, radii, densities)
     gravity = np.interp(radius, radii, gravities)
 
-    liquid_static_derivatives = radial_derivatives_liquid_static(
-        radius, y_vector, density, gravity,
-        order_l=order_l, G_to_use=G_to_use
-        )
+    if incompressible:
+        liquid_static_derivatives = radial_derivatives_liquid_static_incomp(
+            radius, y_vector, density, gravity,
+            order_l=order_l, G_to_use=G_to_use
+            )
+    else:
+        liquid_static_derivatives = radial_derivatives_liquid_static(
+            radius, y_vector, density, gravity,
+            order_l=order_l, G_to_use=G_to_use
+            )
 
     return liquid_static_derivatives