+++ TidalPy/tides/multilayer/numerical_int/derivatives/radial_derivatives_dynamic_incomp.py
@@ -25,13 +25,13 @@ from .....utilities.types import FloatArray, NumArray
 def radial_derivatives_solid_general(
     radius: FloatArray,
     radial_functions: Tuple[NumArray, NumArray, NumArray, NumArray, NumArray, NumArray],
-    shear_modulus: NumArray, bulk_modulus: NumArray, density: FloatArray,
+    shear_modulus: NumArray, density: FloatArray,
     gravity: FloatArray, frequency: FloatArray,
     order_l: int = 2, G_to_use: float = G
     ) -> Tuple[NumArray, NumArray, NumArray, NumArray, NumArray, NumArray]:
     """ Calculates the radial derivative of the radial functions in the most general form - for solid layers.
 
-    Allows for compressibility and dynamic tides.
+    Assumes incompressible and allows for dynamic tides.
     Tidal harmonic l is allowed to be an integer >= 2.
 
     OPT: This and the other radial derivative functions are called many times during integration.
@@ -49,8 +49,6 @@ def radial_derivatives_solid_general(
         Tuple of radial functions for a solid layer (y1, y2, y3, y4, y5, y6)
     shear_modulus : NumArray
         Shear modulus (can be complex for dissipation) at `radius` [Pa]
-    bulk_modulus : NumArray
-        Bulk modulus (can be complex for dissipation) at `radius` [Pa]
     density : FloatArray
         Density at `radius` [kg m-3]
     gravity : FloatArray
@@ -72,15 +70,10 @@ def radial_derivatives_solid_general(
 
     y1, y2, y3, y4, y5, y6 = radial_functions
 
-    # Convert compressibility parameters (the first lame parameter can be complex)
-    lame = (bulk_modulus - (2. / 3.) * shear_modulus)
-
     # Optimizations
     lp1 = order_l + 1.
     lm1 = order_l - 1.
     llp1 = order_l * lp1
-    lame_2mu = lame + 2. * shear_modulus
-    lame_2mu_inverse = 1. / lame_2mu
     r_inverse = 1. / radius
     two_shear_r_inv = 2. * shear_modulus * r_inverse
     density_gravity = density * gravity
@@ -91,19 +84,14 @@ def radial_derivatives_solid_general(
     # See Eq. 82 in TS72 or Eqs. 4--9 in KMN15 or Eqs. 13--18 in B15
     #   Note: There appears to be a missing factor of mu^2 in some of the terms in KMN15.
     # dy2 and dy4 contain all three of: dynamic, viscoelastic, and gravitational terms.
-    dy1 = lame_2mu_inverse * (
-            y1_y3_term * -lame * r_inverse +
-            y2
-    )
+    dy1 = y1_y3_term * -1. * r_inverse
 
     dy2 = r_inverse * (
-            y1 * (dynamic_term - 2. * density_gravity) +
-            y2 * -2. +
+            y1 * (dynamic_term + 12. * shear_modulus * r_inverse - 4. * density_gravity) +
+            y3 * llp1 * (density_gravity - 6. * shear_modulus * r_inverse) +
             y4 * llp1 +
             y5 * density * lp1 +
-            y6 * -density * radius +
-            dy1 * 2. * lame +
-            y1_y3_term * (2. * (lame + shear_modulus) * r_inverse - density_gravity)
+            y6 * -density * radius
     )
 
     dy3 = \
@@ -112,12 +100,11 @@ def radial_derivatives_solid_general(
         y4 * (1. / shear_modulus)
 
     dy4 = r_inverse * (
-            y1 * (density_gravity + two_shear_r_inv) +
-            y3 * (dynamic_term - two_shear_r_inv) +
+            y1 * (density_gravity - 3. * two_shear_r_inv) +
+            y2 * -1. +
+            y3 * (dynamic_term + two_shear_r_inv * (2. * llp1 - 1.))+
             y4 * -3. +
-            y5 * -density +
-            dy1 * -lame +
-            y1_y3_term * -lame_2mu * r_inverse
+            y5 * -density
     )
 
     dy5 = \
@@ -138,13 +125,13 @@ def radial_derivatives_solid_general(
 def radial_derivatives_liquid_general(
     radius: FloatArray,
     radial_functions: Tuple[NumArray, NumArray, NumArray, NumArray],
-    bulk_modulus: NumArray, density: FloatArray,
+    density: FloatArray,
     gravity: FloatArray, frequency: FloatArray,
     order_l: int = 2, G_to_use: float = G
     ) -> Tuple[NumArray, NumArray, NumArray, NumArray]:
     """ Calculates the radial derivative of the radial functions in the most general form - for liquid layers.
 
-    Allows for compressibility and dynamic tides.
+    Assumes incompressible and allows for dynamic tides.
     Tidal harmonic l is allowed to be an integer >= 2.
 
     References
@@ -157,8 +144,6 @@ def radial_derivatives_liquid_general(
         Radius where the radial functions are calculated. [m]
     radial_functions : Tuple[ComplexArray, ComplexArray, ComplexArray, ComplexArray]
         Tuple of radial functions for a solid layer (y1, y2, y5, y6)
-    bulk_modulus : NumArray
-        Bulk modulus (can be complex for dissipation) at `radius` [Pa]
     density : FloatArray
         Density at `radius` [kg m-3]
     gravity : FloatArray
@@ -181,16 +166,10 @@ def radial_derivatives_liquid_general(
     # For the dynamic version, y4 = 0 always in a liquid layer and y3 is defined by y1, y2, and y5 analytically
     y1, y2, y5, y6 = radial_functions
 
-    # Convert compressibility parameters (the first lame parameter can be complex)
-    # For the liquid layer it is assumed that the shear modulus is zero so the lame parameter simply
-    #    equals the bulk modulus
-    lame = bulk_modulus
-
     # Optimizations
     lp1 = order_l + 1.
     lm1 = order_l - 1.
     llp1 = order_l * lp1
-    lame_inverse = 1. / lame
     r_inverse = 1. / radius
     density_gravity = density * gravity
     dynamic_term = -frequency * frequency * density * radius
@@ -198,16 +177,14 @@ def radial_derivatives_liquid_general(
 
     # y3 derivative is undetermined for a liquid layer, but we can calculate its value which is still used in the
     #   other derivatives.
-    y3 = (1. / dynamic_term) * (y2 - density_gravity * y1 + density * y5)
+    y3 = (1. / dynamic_term) * (y2 + density * y5 - density_gravity * y1)
     y1_y3_term = 2. * y1 - llp1 * y3
 
     # Eqs. 11--14 in KMN15 equations look like they don't match TS72 because they applied the rheology already.
     #    and substituted y3.
     # We will use TS72 eq. 87 to allow for a generic rheology and bulk dissipation.
     # # dy2 contain all three of: dynamic, viscoelastic, and gravitational terms.
-    dy1 = \
-        y2 * lame_inverse + \
-        y1_y3_term * -r_inverse
+    dy1 = y1_y3_term * -r_inverse
 
     dy2 = r_inverse * (
             y1 * (dynamic_term - 2. * density_gravity) +